First a approach towards a 2d list in haskell
Haskell takes a 1d array and converts to a 2d list.
Then it does the algorithm - which does only works
on small picture or only with the identity so far.
So it does not work at all.
It converts it back to a 1d array in the end.
TODOA1: Haskell should just (cleverly) work on the
1d array. (Shoud save quite some time).
TODOA2: Maybe rewrite the imgray C function. So it
originally just takes and outputs a 1d array.
TODOB: Can haskell deal with a 2d array input? How?